Added fedora installer script

+++ install_fedora28.sh
@@ -1,4 +1,4 @@
-!/bin/bash
+#!/bin/bash
 
 #Update first
 sudo dnf -y update --refresh
@@ -14,16 +14,28 @@ sudo rm -rf /usr/lib64/libsfml-*
 sudo rm -rf /usr/lib64/pkgconfig/sfml-*
 
 #Basic things
-sudo dnf -y install mc cmake git gtest-devel boost-devel 
-
-#Install SFML
-sudo dnf -y install SFML-devel
-
-cd ~
-rm -rf Graph
-git clone https://github.com/mraggi/Graph
-cd Graph
-mkdir build
-cd build
-cmake ..
-make
+PACKAGES='mc cmake git gtest-devel boost-devel gcc gcc-c++'
+sudo dnf -y install $PACKAGES
+rpm --query --queryformat "" $PACKAGES
+if [ "$?" -eq 0 ] #if there is a problem installing dependencies the script stops
+then
+    #Install SFML
+    sudo dnf -y install SFML-devel
+    PACKAGES='SFML SFML-devel'
+    rpm --query --queryformat "" $PACKAGES
+    if [ "$?" -eq 0 ]
+    then
+    cd ~
+        rm -rf Graph
+        git clone https://github.com/mraggi/Graph
+        cd Graph
+        mkdir build
+        cd build
+        cmake ..
+        make
+    else
+        echo "Error installing SFML"
+    fi
+else
+    echo "Error installing dependencies"
+fi
